+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Better report if rasterized vector files are geospatial ([#1769](../../pull/1769))
 - Provide some latitude in vips multiframe detection ([#1770](../../pull/1770))
+- Don't read multiplane ndpi files with openslide ([#1772](../../pull/1772))
 
 ## 1.30.6